+++ src/api/dataforseo_labs/apple.rs
@@ -72,7 +72,7 @@ pub struct DataForSeoLabsAppleKeywordsForAppRequest {
     pub location_name: Option<String>,
     /// Location code (e.g. `2840`); alternative to `location_name`.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub location_code: Option<i64>,
+    pub location_code: Option<i32>,
     /// Full language name (e.g. `"English"`); alternative to `language_code`.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub language_name: Option<String>,
@@ -87,10 +87,10 @@ pub struct DataForSeoLabsAppleKeywordsForAppRequest {
     pub order_by: Option<Vec<String>>,
     /// Maximum number of returned items (default 100, max 1000).
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit: Option<i64>,
+    pub limit: Option<i32>,
     /// Offset into the result set, for pagination.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub offset: Option<i64>,
+    pub offset: Option<i32>,
     /// User-defined task identifier (max 255 characters).
     #[serde(skip_serializing_if = "Option::is_none")]
     pub tag: Option<String>,
@@ -117,7 +117,7 @@ pub struct DataForSeoLabsAppleAppCompetitorsRequest {
     pub location_name: Option<String>,
     /// Location code (e.g. `2840`); alternative to `location_name`.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub location_code: Option<i64>,
+    pub location_code: Option<i32>,
     /// Full language name (e.g. `"English"`); alternative to `language_code`.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub language_name: Option<String>,
@@ -132,10 +132,10 @@ pub struct DataForSeoLabsAppleAppCompetitorsRequest {
     pub order_by: Option<Vec<String>>,
     /// Maximum number of returned items (default 100, max 1000).
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit: Option<i64>,
+    pub limit: Option<i32>,
     /// Offset into the result set, for pagination.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub offset: Option<i64>,
+    pub offset: Option<i32>,
     /// User-defined task identifier (max 255 characters).
     #[serde(skip_serializing_if = "Option::is_none")]
     pub tag: Option<String>,
@@ -162,7 +162,7 @@ pub struct DataForSeoLabsAppleAppIntersectionRequest {
     pub location_name: Option<String>,
     /// Location code (e.g. `2840`); alternative to `location_name`.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub location_code: Option<i64>,
+    pub location_code: Option<i32>,
     /// Full language name (e.g. `"English"`); alternative to `language_code`.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub language_name: Option<String>,
@@ -177,10 +177,10 @@ pub struct DataForSeoLabsAppleAppIntersectionRequest {
     pub order_by: Option<Vec<String>>,
     /// Maximum number of returned items (default 100, max 1000).
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit: Option<i64>,
+    pub limit: Option<i32>,
     /// Offset into the result set, for pagination.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub offset: Option<i64>,
+    pub offset: Option<i32>,
     /// User-defined task identifier (max 255 characters).
     #[serde(skip_serializing_if = "Option::is_none")]
     pub tag: Option<String>,
@@ -207,7 +207,7 @@ pub struct DataForSeoLabsAppleBulkAppMetricsRequest {
     pub location_name: Option<String>,
     /// Location code (e.g. `2840`); alternative to `location_name`.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub location_code: Option<i64>,
+    pub location_code: Option<i32>,
     /// Full language name (e.g. `"English"`); alternative to `language_code`.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub language_name: Option<String>,
