+++ src/entity/app_data/app_info.rs
@@ -25,7 +25,7 @@ pub struct AppDataApiAppInfoResult {
     /// Total number of results the app store reported.
     pub se_results_count: Option<i64>,
     /// Number of elements returned in `items`.
-    pub items_count: Option<i32>,
+    pub items_count: Option<i64>,
     /// Application info elements returned for the lookup.
     pub items: Option<Vec<AppDataApiAppInfoItem>>,
 }
@@ -38,9 +38,9 @@ pub struct AppDataApiAppInfoItem {
     #[serde(rename = "type")]
     pub item_type: Option<String>,
     /// Position among items of the same `type`.
-    pub rank_group: Option<i32>,
+    pub rank_group: Option<i64>,
     /// Absolute position among all returned items.
-    pub rank_absolute: Option<i32>,
+    pub rank_absolute: Option<i64>,
     /// Alignment of the element within the results (e.g. `left`).
     pub position: Option<String>,
     /// Store identifier of the application.
