+++ src/entity/app_data/app_reviews.rs
@@ -29,7 +29,7 @@ pub struct AppDataApiAppReviewsResult {
     /// Types of result elements present in `items`.
     pub item_types: Option<Vec<String>>,
     /// Number of elements returned in `items`.
-    pub items_count: Option<i32>,
+    pub items_count: Option<i64>,
     /// Reviews returned for the application.
     pub items: Option<Vec<AppDataApiAppReviewsItem>>,
 }
@@ -42,9 +42,9 @@ pub struct AppDataApiAppReviewsItem {
     #[serde(rename = "type")]
     pub item_type: Option<String>,
     /// Position among items of the same `type`.
-    pub rank_group: Option<i32>,
+    pub rank_group: Option<i64>,
     /// Absolute position among all returned items.
-    pub rank_absolute: Option<i32>,
+    pub rank_absolute: Option<i64>,
     /// Alignment of the element within the results (e.g. `left`).
     pub position: Option<String>,
     /// Store identifier of the review.
