+++ src/entity/backlinks/domain_pages_summary.rs
@@ -8,9 +8,9 @@ pub struct BacklinksApiDomainPagesSummary {
     /// Target the metrics refer to (domain, subdomain, or webpage).
     pub target: Option<String>,
     /// Total number of relevant elements in the database.
-    pub total_count: Option<i32>,
+    pub total_count: Option<i64>,
     /// Number of elements in the `items` array.
-    pub items_count: Option<i32>,
+    pub items_count: Option<i64>,
     /// Elements returned for this result.
     pub items: Option<Vec<BacklinksApiElementBacklinksPageSummary>>,
 }
