+++ src/entity/backlinks/element/backlinks.rs
@@ -26,13 +26,13 @@ pub struct BacklinksElementBacklink {
     /// Whether the backlink has been lost.
     pub is_lost: Option<bool>,
     /// Spam score of the backlink (0-100).
-    pub backlink_spam_score: Option<i32>,
+    pub backlink_spam_score: Option<i64>,
     /// Rank of this element on the configured scale (0-1000 by default).
-    pub rank: Option<i32>,
+    pub rank: Option<i64>,
     /// Rank of the referring page.
-    pub page_from_rank: Option<i32>,
+    pub page_from_rank: Option<i64>,
     /// Rank of the referring domain.
-    pub domain_from_rank: Option<i32>,
+    pub domain_from_rank: Option<i64>,
     /// Platform types detected for the referring domain.
     pub domain_from_platform_type: Option<Vec<String>>,
     /// Whether the referring domain is an IP address.
@@ -42,11 +42,11 @@ pub struct BacklinksElementBacklink {
     /// Country of the referring domain.
     pub domain_from_country: Option<String>,
     /// Number of external links on the referring page.
-    pub page_from_external_links: Option<i32>,
+    pub page_from_external_links: Option<i64>,
     /// Number of internal links on the referring page.
-    pub page_from_internal_links: Option<i32>,
+    pub page_from_internal_links: Option<i64>,
     /// Size of the referring page in bytes.
-    pub page_from_size: Option<i32>,
+    pub page_from_size: Option<i64>,
     /// Character encoding of the referring page.
     pub page_from_encoding: Option<String>,
     /// Language of the referring page.
@@ -82,15 +82,15 @@ pub struct BacklinksElementBacklink {
     /// HTML semantic location of the backlink on the referring page.
     pub semantic_location: Option<String>,
     /// Number of links from the referring page to the target.
-    pub links_count: Option<i32>,
+    pub links_count: Option<i64>,
     /// Number of grouped backlinks represented by this item.
-    pub group_count: Option<i32>,
+    pub group_count: Option<i64>,
     /// Whether the backlink is broken.
     pub is_broken: Option<bool>,
     /// HTTP status code returned by the `url_to` page.
     pub url_to_status_code: Option<i32>,
     /// Spam score of the `url_to` page (0-100).
-    pub url_to_spam_score: Option<i32>,
+    pub url_to_spam_score: Option<i64>,
     /// Redirect target of the `url_to` page, if any.
     pub url_to_redirect_target: Option<String>,
     /// Ranked-keyword counts for the referring page.
@@ -105,11 +105,11 @@ pub struct BacklinksElementBacklink {
 #[derive(Debug, Default, Serialize, Deserialize, Clone)]
 pub struct BacklinksElementBacklinkRankKeywordsInfo {
     /// Number of keywords the referring page ranks for in the top 3.
-    pub page_from_keywords_count_top_3: Option<i32>,
+    pub page_from_keywords_count_top_3: Option<i64>,
     /// Number of keywords the referring page ranks for in the top 10.
-    pub page_from_keywords_count_top_10: Option<i32>,
+    pub page_from_keywords_count_top_10: Option<i64>,
     /// Number of keywords the referring page ranks for in the top 100.
-    pub page_from_keywords_count_top_100: Option<i32>,
+    pub page_from_keywords_count_top_100: Option<i64>,
 }
 
 /// A single hop in the redirect/canonical chain of an indirect backlink.
