+++ src/entity/backlinks/element/backlinks_bulk_new_lost_referring_domains.rs
@@ -10,11 +10,11 @@ pub struct BacklinksApiElementBulkNewLostReferringDomains {
     /// Target the metrics refer to (domain, subdomain, or webpage).
     pub target: Option<String>,
     /// Number of newly discovered referring domains.
-    pub new_referring_domains: Option<i32>,
+    pub new_referring_domains: Option<i64>,
     /// Number of referring domains that were lost.
-    pub lost_referring_domains: Option<i32>,
+    pub lost_referring_domains: Option<i64>,
     /// Number of newly discovered referring main domains.
-    pub new_referring_main_domains: Option<i32>,
+    pub new_referring_main_domains: Option<i64>,
     /// Number of referring main domains that were lost.
-    pub lost_referring_main_domains: Option<i32>,
+    pub lost_referring_main_domains: Option<i64>,
 }
